Refactor: split operators.py into an operators/ package
operators.py had grown to 849 LOC holding the QuadraticForm protocol,
shared solver primitives, and four covariance backends. Split it into a
package so no single module dominates and each stays well under 500 LOC,
with no change to the public cvxcla.operators import surface:

- _core.py   QuadraticForm protocol, _rcond_symmetric, bordered_solve,
             CovarianceOperator alias
- dense.py   DenseCovariance, IncrementalDenseCovariance
- factor.py  FactorCovariance
- gram.py    GramCovariance
- __init__.py re-exports the public names (same imports keep working)

Behaviour-preserving: docstrings and code moved verbatim. All gates green
(ruff, ty + mypy --strict, interrogate 100%, deptry, pip-audit/bandit,
validate, 243 tests, 100% coverage).
